feat: Implement repositories for payment methods, roles, users, wallets, and webhooks
- Added IPaymentMethodRepository with methods for managing user payment methods.
- Introduced IRoleRepository for role management functionalities.
- Created IUserRepository for user data handling, including personal data and roles.
- Developed IWalletRepository for wallet operations, including account and transaction management.
- Implemented IWebhookRepository for handling webhook events.
- Established models for user payment methods: UserCard, UserBankAccount, UserPixKey.
- Updated Order model to include additional payment-related fields.
- Refactored LedgerEntry and PaymentIntent models to use decimal for amounts.
- Added repository implementations for all new interfaces.
- Removed obsolete PaymentMethod model.

+++ src/Product.Contracts/Wallet/AmountRequest.cs
@@ -2,5 +2,5 @@ namespace Product.Contracts.Wallet;
 
 public class AmountRequest
 {
-    public long Amount { get; set; }
+    public decimal Amount { get; set; }
 }
