+++ src/Product.Data/Models/Wallet/PaymentIntent.cs
@@ -8,7 +8,7 @@ public class PaymentIntent : Entity<Guid>
 {
     public Guid UserId { get; set; }
     public string Provider { get; set; } = null!;
-    public long Amount { get; set; }
+    public decimal Amount { get; set; }
     public string Currency { get; set; } = "BRL";
     public PaymentIntentStatus Status { get; set; } = PaymentIntentStatus.PENDING;
     public string? ExternalPaymentId { get; set; }
